pyproject.toml
[project]
name = "autosec-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openrouter>=0.2.1",
    "load-dotenv>=0.1.0",
    "openai>=2.31.0",
]
